test: consolidate gateway test files, remove duplicate coverage file
Merges test_gateway_coverage.py (57 duplicate + 23 unique classes) into test_gateway_helpers.py. The 23 unique classes are preserved; the 2,139-line duplicate file is deleted. No test logic changed.
